Feat: featuring RestClient for get cloth recommendation form fastAPI server

+++ ollive_back/src/main/java/org/palpalmans/ollive_back/domain/cloth/model/dto/request/ClothRecommendationRequest.java
@@ -2,7 +2,7 @@ package org.palpalmans.ollive_back.domain.cloth.model.dto.request;
 
 import java.util.Objects;
 
-public record ClothRecommendationRequest(String text, int longitude, int latitude) {
+public record ClothRecommendationRequest(String text, double longitude, double latitude) {
     public ClothRecommendationRequest {
         Objects.requireNonNull(text);
     }
